refactor: member 도메인 패키지 세분화
