+++ src/main/java/org/ject/support/common/security/CustomUserDetailService.java
@@ -1,15 +1,15 @@
 package org.ject.support.common.security;
 
 
-import static org.ject.support.domain.member.MemberErrorCode.NOT_FOUND_MEMBER;
-
 import lombok.RequiredArgsConstructor;
-import org.ject.support.domain.member.Member;
-import org.ject.support.domain.member.MemberException;
-import org.ject.support.domain.member.MemberRepository;
+import org.ject.support.domain.member.entity.Member;
+import org.ject.support.domain.member.exception.MemberException;
+import org.ject.support.domain.member.repository.MemberRepository;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+
+import static org.ject.support.domain.member.exception.MemberErrorCode.NOT_FOUND_MEMBER;
 
 @Service
 @RequiredArgsConstructor
