+++ src/main/java/org/ject/support/domain/member/entity/Team.java
@@ -1,6 +1,10 @@
-package org.ject.support.domain.member;
+package org.ject.support.domain.member.entity;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
