+++ src/main/java/org/ject/support/domain/member/entity/TeamMember.java
@@ -1,6 +1,12 @@
-package org.ject.support.domain.member;
+package org.ject.support.domain.member.entity;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.ject.support.domain.base.BaseTimeEntity;
