+++ src/main/java/org/ject/support/domain/member/repository/MemberRepository.java
@@ -1,6 +1,7 @@
-package org.ject.support.domain.member;
+package org.ject.support.domain.member.repository;
 
 import java.util.Optional;
+import org.ject.support.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
