+++ src/main/java/org/ject/support/domain/member/repository/TeamRepository.java
@@ -1,5 +1,6 @@
-package org.ject.support.domain.member;
+package org.ject.support.domain.member.repository;
 
+import org.ject.support.domain.member.entity.Team;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TeamRepository extends JpaRepository<Team, Long> {
