+++ src/main/java/org/ject/support/domain/project/entity/Project.java
@@ -1,11 +1,21 @@
 package org.ject.support.domain.project.entity;
 
-import jakarta.persistence.*;
-import lombok.*;
-import org.ject.support.domain.base.BaseTimeEntity;
-import org.ject.support.domain.member.Team;
-
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import java.time.LocalDate;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.ject.support.domain.base.BaseTimeEntity;
+import org.ject.support.domain.member.entity.Team;
 
 @Entity
 @Getter
