+++ src/main/java/org/ject/support/domain/recruit/domain/ApplicationForm.java
@@ -1,10 +1,17 @@
 package org.ject.support.domain.recruit.domain;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.ject.support.domain.base.BaseTimeEntity;
-import org.ject.support.domain.member.Member;
+import org.ject.support.domain.member.entity.Member;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
